notebook/kg_ai_alg/chatbot/neptune-genai-examples/llamaindex/knowledgegraphindex-chatbot-streamlit/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.20.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.76
botocore==1.34.76
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
cssselect==1.2.0
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
jieba3k==0.35.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llama-index-core>=0.10.26
llama-index-embeddings-bedrock>=0.1.4
llama-index-graph-stores-neptune>=0.1.3
llama-index-llms-anthropic>=0.1.7
llama-index-llms-bedrock>=0.1.5
llama-index-readers-web>=0.1.7
llamaindex-py-client>=0.1.15
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.16.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
playwright==1.42.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
selenium==4.19.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
streamlit==1.32.2
tenacity==8.2.3
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
llama-index-vector-stores-neptune>=0.1.0
